Struct rust_gpu_builder::ShaderBuilder

pub(crate) struct ShaderBuilder {

    pub(crate) path_to_crate: PathBuf,
    pub(crate) output_path: Option<PathBuf>,
    pub(crate) output_format: OutputFormat,
    pub(crate) target: String,
    pub(crate) deny_warnings: bool,
    pub(crate) release: bool,
    pub(crate) capability: Vec<Capability>,
    pub(crate) multimodule: bool,
    pub(crate) spirv_metadata: SpirvMetadata,
    pub(crate) relax_struct_store: bool,
    pub(crate) relax_logical_pointer: bool,
    pub(crate) relax_block_layout: bool,
    pub(crate) uniform_buffer_standard_layout: bool,
    pub(crate) scalar_block_layout: bool,
    pub(crate) skip_block_layout: bool,
    pub(crate) preserve_bindings: bool,
    pub(crate) watch_paths: Option<Vec<String>>,

}

Clap application struct.

Fields
path_to_crate: PathBuf

Shader crate to compile.

output_path: Option<PathBuf>

If set, combined SPIR-V and entrypoint metadata will be written to this file on succesful compile.

output_format: OutputFormat

The format to write output in.

target: String

rust-gpu compile target.

deny_warnings: bool

Treat warnings as errors during compilation.

release: bool

Compile shaders in release mode.

capability: Vec<Capability>

Enables the provided SPIR-V capability.

multimodule: bool

Compile one .spv file per entry point.

spirv_metadata: SpirvMetadata

Set the level of metadata included in the SPIR-V binary.

relax_struct_store: bool

Allow store from one struct type to a different type with compatible layout and members.

relax_logical_pointer: bool

Allow allocating an object of a pointer type and returning a pointer value from a function in logical addressing mode.

relax_block_layout: bool

Enable VK_KHR_relaxed_block_layout when checking standard uniform, storage buffer, and push constant layouts. This is the default when targeting Vulkan 1.1 or later.

uniform_buffer_standard_layout: bool

Enable VK_KHR_uniform_buffer_standard_layout when checking standard uniform buffer layouts.

scalar_block_layout: bool

Enable VK_EXT_scalar_block_layout when checking standard uniform, storage buffer, and push constant layouts. Scalar layout rules are more permissive than relaxed block layout so in effect this will override the –relax-block-layout option.

skip_block_layout: bool

Skip checking standard uniform / storage buffer layout. Overrides any –relax-block-layout or –scalar-block-layout option.

preserve_bindings: bool

Preserve unused descriptor bindings. Useful for reflection.

watch_paths: Option<Vec<String>>

If set, will watch the provided directory and recompile on change.

Can be specified multiple times to watch more than one directory.

Implementations

impl ShaderBuilder

pub(crate) fn spirv_metadata(s: &str) -> Result<SpirvMetadata, Error>

Clap value parser for SpirvMetadata.

pub(crate) fn spirv_capability(s: &str) -> Result<Capability, Error>

Clap value parser for Capability.

pub fn build_shader(&self) -> Result<CompileResult, SpirvBuilderError>

Builds a shader with the provided set of options.